Hide memory-efficient-attention behind a feature flag
